% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fangbian-package.R
\docType{package}
\name{fangbian-package}
\alias{fangbian}
\alias{fangbian-package}
\title{fangbian: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
