% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logi.R
\name{logi}
\alias{logi}
\title{拟合逻辑回归以及提取模型结果}
\usage{
logi(
  data,
  y,
  x,
  covars = NULL,
  q_grp = 4,
  q_val = NULL,
  q_ref = 1,
  digit_grp = 2,
  digit_rr = 2,
  log_info = TRUE
)
}
\arguments{
\item{data}{数据集（必须是数据框或tibble）}

\item{y}{结局变量名称（字符，长度为1，必须是二分类）}

\item{x}{感兴趣的自变量名称（字符型，长度为1，x对应的数据必须是连续型）}

\item{covars}{协变量名称（字符型，长度大于等于0，可以为空）}

\item{q_grp}{分组数，将x对应的连续型数据划分为几组（正整数，长度为1），不能与q_val同时设置}

\item{q_val}{分组的界值，将x划分为不同的组别（数值型，长度任意），不能与q_grp同时设置}

\item{q_ref}{参考组，指定x转换为分类变量后，第几组为参考组（正整数，长度为1，默认值为1）}

\item{digit_grp}{数值保留小数位数，用于分组范围的展示（正整数，默认值为2）}

\item{digit_rr}{rr值及95\%CI保留的小数位数（正整数，默认值为2）}

\item{log_info}{逻辑型，是否显示日志，默认为TRUE}
}
\value{
返回一个列表，列表中有4个元素，具体如下：

\strong{元素1: raw (type = list)}
\itemize{
\item raw_model1: 模型1（原始连续型变量x）原始结果 (type = summary.glm)
\item raw_model2: 模型2（变量x标化后）原始结果 (type = summary.glm)
\item raw_model3: 模型3（变量x分类后）原始结果 (type = summary.glm)
\item raw_model4: 模型4（变量x分类后，再转化为数值，仅用于提取趋势性检验P值）原始结果 (type = summary.glm)
}

\strong{元素2: model1_continuous (type = data.frame)}
\itemize{
\item 模型1提取后的结果
}

\strong{元素3: model2_per_SD (type = data.frame)}
\itemize{
\item 模型2提取后的结果
}

\strong{元素4: model3_classified  (type = data.frame)}
\itemize{
\item 模型3提取后的结果（相对于前2个模型，多了一个趋势性检验P值）
}

\strong{元素5: group (type = data.frame)}
\itemize{
\item 模型3中变量x的分组标准
}

model1_continuous, model2_per_SD, model3_classified中变量定义如下：
\itemize{
\item term: 连续型变量x的名称
\item rr: RR值 (原始)
\item lo: RR值的95\%CI下限 (原始)
\item hi: RR值的95\%CI上限 (原始)
\item rr2: RR值 (保留2位小数)
\item lo2: RR值的95\%CI下限 (保留2位小数)
\item hi2: RR值的95\%CI上限 (保留2位小数)
\item ci: RR值和95\%CI的组合：rr (lo, hi)
\item p: RR值对应的P值 (原始)
\item p2: 如果p>0.05则显示保留2位小数的p值，如果p<0.001则显示"<0.001"，，如果p<0.05则显示"<0.05"
\item PforTrend: 趋势性检验P值（只有model3_classified才有这个变量）
\item PforTrend2: 计算方法同p2
}

group中的变量定义如下：
\itemize{
\item group_var: 变量x的名称
\item group_name: 变量x分组后的名称
\item group_range: 每组的取值范围
\item group_ref: 哪一组是参考组 (Ref)
}
}
\description{
该函数将拟合3种Logistics回归，并提取模型结果。
第一种是将变量x(连续型)直接纳入模型；
第二种是将变量x标化后(mean = 0, sd = 1)纳入模型；
第三种是将变量x按照用户指定的分组方式(默认为分位数分组)转化为分类变量后纳入模型；
}
\examples{
# 示例数据生成
set.seed(123)
test_data <- data.frame(
  is_bapwv = sample(c(0, 1), 100, replace = TRUE),  # 二分类变量
  sbp = rnorm(100),  # 连续变量
  covar1 = rnorm(100),
  covar2 = sample(c("A", "B"), 100, replace = TRUE)
)

# 调用 logi1 函数拟合逻辑回归模型
output <- logi(
  data = test_data,
  y = "is_bapwv",
  x = "sbp",
  covars = c("covar1", "covar2"),
  q_grp = 4
)

# 输出结果
print(output)
}
