% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logi1_sub.R
\name{logi1_sub}
\alias{logi1_sub}
\title{亚组Logistics模型拟合与结果提取}
\usage{
logi1_sub(data, y, x, covar = NULL, by, log_info = TRUE)
}
\arguments{
\item{data}{数据集}

\item{y}{结局变量的名字，字符型}

\item{x}{感兴趣的自变量x的名字，字符型}

\item{covar}{协变量的名字，字符型，可以有多个，也可以没有}

\item{by}{分组变量的名字，字符型，可以有多个}

\item{log_info}{逻辑型，是否显示日志，默认为TRUE}
}
\value{
一个包含原始模型结果、交互作用P值结果、提取结果的列表:

raw: 一个嵌套的列表，包含所有分组变量及其水平的原始模型结果
\itemize{
\item 每个分组变量对应一个列表，例如 results_raw$age，包含该分组变量各个水平的原始模型结果
\item 每个水平中的列表包含通过 summary() 函数获得的模型摘要，包括系数估计、标准误、z 值和 P 值等
}

PforInteract: 一个列表，包含每个分组变量的交互作用 P 值结果
\itemize{
\item 每个分组变量的名称作为列表的键，例如 results_pforinteract$age
\item 每个键对应的值是通过 anova() 计算的交互作用 P 值，用于判断分组变量与感兴趣自变量之间是否存在显著交互作用
}

extracted_results: 一个 tibble，存储提取后的主要结果，方便进一步分析或呈现
\itemize{
\item y: 结局变量的名称
\item x: 感兴趣的自变量名称
\item subset: 分组变量的名称
\item level: 分组变量的具体水平
\item sample: 当前分组水平中的总样本数量
\item cases: 当前分组水平中结局事件为 1 的样本数量
\item PforInteraction: 交互作用 P 值，对应于该分组变量是否有显著交互
\item rr: 相对风险 (RR) 值，即通过对自变量进行逻辑回归模型拟合后计算得到的指数化系数
\item lo 和 hi: 相对风险的 95\% 置信区间下限和上限，表示估计值的不确定性范围
}
}
\description{
logi1_sub 用于在分组的基础上拟合逻辑回归模型，
帮助用户分析感兴趣的自变量与结局变量之间的关系，
并根据分组变量逐层提取结果。
函数能够对每个分组的不同水平分别拟合模型，
并提取回归系数、相对风险（RR）及其置信区间等信息。
此外，函数还会计算各分组变量与自变量之间的交互作用 P 值，
以评估自变量与分组变量之间的潜在交互效果。
}
\examples{
set.seed(123)
test_data <- data.frame(
  is_bapwv = sample(c(0, 1), 100, replace = TRUE),  # 二分类变量
  sbp = rnorm(100),  # 连续变量
  covar1 = rnorm(100),
  covar2 = rnorm(100),
  covar3 = rnorm(100),
  type = sample(c("A", "B"), 100, replace = TRUE),
  sex =  sample(c("f", "m"), 100, replace = TRUE)
)
a = logi1_sub(data = test_data,
y = 'is_bapwv',
x = 'sbp', covar = c('covar1', 'covar2', 'covar3'),
by = c('type', 'sex'),
log_info = T)
print(a)
}
