% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logi1.R
\name{logi1}
\alias{logi1}
\title{拟合逻辑回归以及提取模型结果}
\usage{
logi1(data, y, x, covars, q_grp = NULL, q_val = NULL, q_ref = 1, digit_grp = 2)
}
\arguments{
\item{data}{数据集（必须是数据框或tibble）}

\item{y}{结局变量名称（字符，长度为1，必须是二分类）}

\item{x}{感兴趣的自变量名称（字符型，长度为1，x对应的数据必须是连续型）}

\item{covars}{协变量名称（字符型，长度大于等于0，可以为空）}

\item{q_grp}{分组数，将x对应的连续型数据划分为几组（正整数，长度为1），不能与q_val同时设置}

\item{q_val}{分组的界值，将x划分为不同的组别（数值型，长度任意），不能与q_grp同时设置}

\item{q_ref}{参考组，指定x转换为分类变量后，第几组为参考组（正整数，长度为1，默认值为1）}

\item{digit_grp}{数值保留小数位数，用于分组范围的展示（正整数，默认值为2）}
}
\value{
包含模型摘要和其他信息的列表。
}
\description{
该函数适合逻辑回归模型并提取关键结果。
}
\examples{
# 示例数据生成
set.seed(123)
test_data <- data.frame(
  is_bapwv = sample(c(0, 1), 100, replace = TRUE),  # 二分类变量
  sbp = rnorm(100),  # 连续变量
  covar1 = rnorm(100),
  covar2 = sample(c("A", "B"), 100, replace = TRUE)
)

# 调用 logi1 函数拟合逻辑回归模型
output <- logi1(
  data = test_data,
  y = "is_bapwv",
  x = "sbp",
  covars = c("covar1", "covar2"),
  q_grp = 4
)

# 输出结果
print(output)
}
